Add support for EC2 measurement
EC2 launched SEV-SNP support on 2023-04-28 [1]. This patch adds support
to regenerate measurements for SEV-SNP guests running in EC2.

Today, sev-snp-measure always expects QEMU's initialization logic and
order. But EC2's measurement calculation is not 100% identical with QEMU's.
In EC2, we set up tiny bits and pieces of initial register and memory
state up differently.

This patch introduces a way to pick different VMM types: 'QEMU' and
'ec2'. It further enlightens all logic that differs between the two
to use the EC2 mechanism when the VMM type is ec2.

[1] https://aws.amazon.com/about-aws/whats-new/2023/04/amazon-ec2-amd-sev-snp/

Signed-off-by: Alexander Graf <graf@amazon.com>