Add support for OVMF hashes
Some times we may not want to have to keep the full OVMF binary handy
when we try to validate a measurement. Most of the data we extract out
of the OVMF binary is static, except for its actual hashes.

So let's introduce a new mode that generates a precalculated launch
digest that contains the SNP ld hash at the point where we fully
ingested OVMF's contents:

  $ sev-snp-measure --mode snp:ovmf-hash --ovmf OVMF.fd
  cab7e[...]

In addition, add a new optional parameter that a user can then use to
consume the hash instead of recalculating it from their OVMF binary at
hand:

  $ sev-snp-measure --mode snp --vcpus=1 --vcpu-type=EPYC-v4 \
                    --ovmf=OVMF.fd.old --snp-ovmf-hash cab7e[...]
  d5269[...]

which is identical to the full calculation with the new OVMF.fd:

  $ sev-snp-measure --mode snp --vcpus=1 --vcpu-type=EPYC-v4 --ovmf=OVMF.fd
  d5269[...]

With this in place, we no longer need to copy full OVMF binaries around.
Instead, for almost all OVMF binaries in existence today, we can merely
share their hash values and are still able to validate the measurement
correctness.

Signed-off-by: Alexander Graf <graf@amazon.com>