Add SNP:SVSM mode
Add a mode to calculate the SNP launch measurement for the SNP+Coconut
SVSM use-case.

A new mode is added "snp:svsm". It currently only works for Qemu.
If selected, two new command line arguments need to be set as well:

`--svsm` points to he svsm binary. This is to be included in the
measurement.

`--vars-size` is the size in bytes of the OVMF variable store file. This
is required to calculate the offset at which the SVSM will be located in
the guest memory.

See https://github.com/coconut-svsm/svsm

Signed-off-by: Oliver Steffen <osteffen@redhat.com>