cli: rework snp:svsm mode arguments and help
Add an additional argument to the snp:svsm mode: --vars-file.
This can be used to specify the OVMF_VARS file directly, instead of
getting the file size and using --vars-size.
The option is mutually exclusive with --vars-size.

Rework the argument parsing for snp:svsm mode, using an argument group
for this mode. The original behavior is not changed.

Make the help text more clear.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>